# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: GBF ERP API
  version: "0.2.0"
  description: |
    通用贸易 MVP 后端 HTTP 接口（**精简说明**）。完整行为以 `src/routes` 与 Zod schema 为准。

    - 成功响应多为 `application/json`；错误体由全局错误处理统一包装。
    - 鉴权：除 `/health`、`/api/auth/login`、`/api/auth/refresh`（及开发态 `x-username`）外，一般需 `Authorization: Bearer <access_token>`。
servers:
  - url: http://localhost:3100
    description: 本地默认端口
...
paths:
...
  /api/suppliers:
    get:
      tags: [master]
      summary: 供应商列表
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
    post:
      tags: [master]
      summary: 创建供应商
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PartnerBody" }
      responses:
        "201":
          description: Created
  /api/customers:
    get:
      tags: [master]
      summary: 客户列表
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
    post:
      tags: [master]
      summary: 创建客户
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PartnerBody" }
      responses:
        "201":
          description: Created
  /api/products:
    get:
      tags: [master]
      summary: 商品列表（含库存汇总）
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
    post:
      tags: [master]
      summary: 创建商品（需 stock:write 或 product:write）
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductBody" }
      responses:
        "201":
          description: Created
  /api/stock/ledger:
    get:
      tags: [master]
      summary: 库存流水
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
...
  /api/ar/invoices:
    get:
      tags: [finance]
      summary: 应收发票列表
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
  /api/ap/bills:
    get:
      tags: [finance]
      summary: 应付账单列表
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
  /api/finance/receipts:
    post:
      tags: [finance]
      summary: 创建收款单
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReceiptBody" }
      responses:
        "201":
          description: Created
  /api/finance/payments:
    post:
      tags: [finance]
      summary: 创建付款单
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PaymentBody" }
      responses:
        "201":
          description: Created
  /api/finance/accounts:
    get:
      tags: [finance]
      summary: 会计科目
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
  /api/finance/journals:
    get:
      tags: [finance]
      summary: 凭证/日记账列表
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: access token（`typ=access`）
...
  schemas:
...
    PartnerBody:
      type: object
      required: [code, name]
      properties:
        code: { type: string }
        name: { type: string }
        contact: { type: string }
        phone: { type: string }
    ProductBody:
      type: object
      required: [sku, name, unit, costPrice, salePrice]
      properties:
        sku: { type: string, minLength: 2 }
        name: { type: string, minLength: 2 }
        unit: { type: string, minLength: 1 }
        costPrice: { type: number, minimum: 0 }
        salePrice: { type: number, minimum: 0 }
...
    ReceiptBody:
      type: object
      required: [receiptNo, customerId, amount]
      properties:
        receiptNo: { type: string, minLength: 3 }
        customerId: { type: integer, minimum: 1 }
        arInvoiceId: { type: integer, minimum: 1 }
        amount: { type: number, minimum: 0 }
    PaymentBody:
      type: object
      required: [paymentNo, supplierId, amount]
      properties:
        paymentNo: { type: string, minLength: 3 }
        supplierId: { type: integer, minimum: 1 }
        apBillId: { type: integer, minimum: 1 }
        amount: { type: number, minimum: 0 }